requirements.txt
asyncio-mqtt
paho-mqtt
python-dotenv
numpy
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints
